requirements.txt
fastapi==0.104.1
uvicorn==0.29.0
sqlalchemy==2.0.23
python-dotenv==1.0.1
psycopg2-binary==2.9.9